Refactor codebase with proper separation of concerns
Major architectural improvements for maintainability and scalability:

## Shared Utilities (New)
- src/shared/constants.js - Centralized constants and configurations
- src/shared/storage.js - Chrome storage wrapper with type-safe functions

## Content Script Refactoring
Split 377-line index.js into modular components:
- src/content/ui/modal.js - Modal UI creation and DOM manipulation
- src/content/ui/toast.js - Toast notification system
- src/content/services/savings.js - Savings tracking business logic
- src/content/services/reminders.js - Reminder creation service
- src/content/utils/price.js - Price extraction utilities
- src/content/index.js - Main orchestrator (187 lines, down from 377)
- src/content/checkout.js - Updated to use shared constants

## Background Script Refactoring
Split 336-line index.js into modular components:
- src/background/config/keys.js - API key management and rotation
- src/background/api/prompts.js - AI prompt builders for personalities
- src/background/api/openrouter.js - OpenRouter API client and streaming
- src/background/services/debate.js - Debate orchestration logic
- src/background/services/reminders.js - Reminder checking and notifications
- src/background/index.js - Main orchestrator (78 lines, down from 336)

## Popup & Options Updates
- Updated to use shared constants and storage wrappers
- Consistent use of MESSAGE_TYPES constants
- Type-safe storage operations

## Benefits
- Single Responsibility Principle - Each module has one clear purpose
- Easier testing - Can test individual modules in isolation
- Better maintainability - Clear code organization
- Improved readability - Smaller, focused files
- Easier to extend - New features can be added without touching core logic
- Reduced duplication - Shared utilities eliminate repeated code

## Metrics
- Content script: 377 lines → 187 lines (main orchestrator)
- Background script: 336 lines → 78 lines (main orchestrator)
- Total new modules: 11 files
- Build successful with all imports resolved
