+++ src/options/controller.js
@@ -1,17 +1,14 @@
 // Options Page Controller
 
-const DEFAULT_MODELS = {
-  enabler: 'anthropic/claude-3-haiku',
-  skeptic: 'anthropic/claude-3.5-sonnet',
-  mediator: 'anthropic/claude-3-opus'
-};
+import { DEFAULT_MODELS, DEFAULT_PRICE_THRESHOLD } from '../shared/constants.js';
+import { getSyncStorage, setSyncStorage, clearSyncStorage } from '../shared/storage.js';
 
 // Load saved settings
 async function loadSettings() {
-  const result = await chrome.storage.sync.get(['apiKey', 'models', 'priceThreshold']);
+  const result = await getSyncStorage(['apiKey', 'models', 'priceThreshold']);
 
   document.getElementById('apiKey').value = result.apiKey || '';
-  document.getElementById('priceThreshold').value = result.priceThreshold !== undefined ? result.priceThreshold : 50;
+  document.getElementById('priceThreshold').value = result.priceThreshold !== undefined ? result.priceThreshold : DEFAULT_PRICE_THRESHOLD;
 
   // Load personality-specific models
   const models = result.models || DEFAULT_MODELS;
@@ -45,7 +42,7 @@ async function saveSettings(e) {
   }
 
   // Save to Chrome storage
-  await chrome.storage.sync.set({ apiKey, models, priceThreshold });
+  await setSyncStorage({ apiKey, models, priceThreshold });
 
   showStatus('Settings saved successfully!', 'success');
 }
@@ -56,7 +53,7 @@ async function resetSettings() {
     return;
   }
 
-  await chrome.storage.sync.clear();
+  await clearSyncStorage();
   await loadSettings();
 
   showStatus('Settings reset to defaults', 'success');
